Add reset collection step (closes #11)
This commit introduces a new 'unindexed' step that triggers a reindex step which resets the entire NFT collection from scratch. This is a little different from the current indexer implementation in Node.js, where the `CreationFlow.CollectionCreator` step triggers this action instead. I chose to separate these to make a clear distinction between those two steps, for a better separation of concerns.